# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/blog/ideas:
    post:
      operationId: generateBlogIdeas
      tags: [admin-blog]
      summary: Generate post-title ideas for a macrocategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogIdeasInput"
      responses:
        "200":
          description: Generated title ideas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogIdeasResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /admin/blog/posts:
    get:
      operationId: listAdminPosts
      tags: [admin-blog]
      summary: List all generated posts
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      operationId: createBlogPost
      tags: [admin-blog]
      summary: Generate a draft blog post (not published)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created draft post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /admin/blog/posts/{id}:
    patch:
      operationId: updateBlogPost
      tags: [admin-blog]
      summary: Update a post (edit content, publish or unpublish)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBlogPostInput"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      operationId: deleteBlogPost
      tags: [admin-blog]
      summary: Delete a generated post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    BlogPostSection:
      type: object
      properties:
        heading:
          type: ["string", "null"]
        paragraphs:
          type: array
          items:
            type: string
      required:
        - paragraphs
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        category:
          type: string
        title:
          type: string
        subtitle:
          type: string
        excerpt:
          type: string
        keywords:
          type: array
          items:
            type: string
        readingMinutes:
          type: integer
        body:
          type: array
          items:
            $ref: "#/components/schemas/BlogPostSection"
        bodyHtml:
          type: string
        oabClosing:
          type: string
        published:
          type: boolean
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - category
        - title
        - subtitle
        - excerpt
        - keywords
        - readingMinutes
        - body
        - bodyHtml
        - oabClosing
        - published
        - createdAt
    BlogIdeasInput:
      type: object
      properties:
        category:
          type: string
          minLength: 1
      required:
        - category
    BlogIdeasResult:
      type: object
      properties:
        ideas:
          type: array
          items:
            type: string
      required:
        - ideas
    BlogPostInput:
      type: object
      properties:
        category:
          type: string
          minLength: 1
        theme:
          type: string
          minLength: 1
      required:
        - category
        - theme
    UpdateBlogPostInput:
      type: object
      properties:
        category:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        subtitle:
          type: string
        excerpt:
          type: string
        bodyHtml:
          type: string
        oabClosing:
          type: string
        published:
          type: boolean
      additionalProperties: false
...